Rename TimelineView to Timeline
