+++ src/frontend/views/Timeline/Clip.styl
