+++ src/frontend/views/Timeline/LaneLabel.i18n.ts
