+++ src/frontend/views/Timeline/_Clip.i18n.ts
