+++ src/frontend/views/Timeline/_Clip.tsx
