+++ src/frontend/views/Timeline/_Gradations.i18n.tsx
