+++ src/frontend/views/Timeline/_Gradations.tsx
