--- src/frontend/views/TimelineView/package.json
@@ -1,4 +1,0 @@
-{
-    "main": "./TimelaneView.tsx",
-    "typings": "./TimelaneView.tsx"
-}
